js/xpconnect/tests/mochitest/test_bug691059.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=691059
-->
<head>
  <title>Test for Bug 691059</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=691059">Mozilla Bug 691059</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">


/** Test for Bug 691059 **/

function f() {}

function testEventTarget(obj) {
  obj.onmouseenter = f;
  is(obj.onmouseenter, f,
     "onmouseenter should be settable");
  obj.onmouseleave = f;
  is(obj.onmouseleave, f,
     "onmouseenter should be settable");
}

function testInterface(obj) {
  try {
    obj.prototype.onmouseenter = f;
    is(obj.onmouseenter, null,
       "prototype.onmouseenter should be settable, but the value should stay null");
    obj.prototype.onmouseleave = f;
    is(obj.onmouseleave, null,
       "prototype.onmouseleave should be settable, but the value should stay null");
  } catch(ex) {
    ok(false, ex);
  }
}

testEventTarget(window);
testEventTarget(document);
testEventTarget(document.documentElement);

testInterface(Document);
testInterface(HTMLElement);

</script>
</pre>
</body>
</html>
